플레이어 키입력 처리(임시)
-키입력시 해당 방향 바라보게함
-움직이는 오브젝트 (추가예정)

+++ RPG_MyLife/Component.cpp
@@ -23,7 +23,7 @@ void Component::ReciverMessage(MessageFrom msgFrom)
 	}
 }
 
-void Component::Init(int posX,int posY)
+void Component::Init(float posX,float posY)
 {
 	_posX = posX;
 	_posY = posY;
